requirements.txt
openai>=0.11.2
requests==2.28.2
streamlit==1.21.0
python-dotenv==1.0.0
pygithub